project/requirements.txt
pandas
numpy
scipy
scikit-learn
river
evidently
pydantic
kafka-python
fastapi
uvicorn
requests
joblib
pyyaml
